Revert "working mamba for training"
This reverts commit f0a4f9369a2883cf5d3b4739e65941f44a8292b6.

+++ hswvime_ppo/buffer.py
@@ -112,7 +112,7 @@ class TransitionRolloutBuffer(RolloutBuffer):
         )
         self.questions = np.zeros(
             (self.buffer_size, self.n_envs, self.num_qa),
-            dtype=np.float64,
+            dtype=np.long,
         )
         self.answers = np.zeros(
             (self.buffer_size, self.n_envs, self.num_qa, self.answer_dim),
